add checkboxes to display and set action status
